move grid into client comp, add suspense boundaries
